spotify_top_songs/requirements.txt
Flask~=3.0.2
oauth2client~=4.1.3
google-api-python-client~=1.7.12
requests~=2.31.0
python-dotenv~=1.0.1